질문 기록 insert 로직 1차 커밋

+++ src/main/java/com/AI/chatbot/repository/QuestionRepository.java
@@ -1,7 +1,9 @@
 package com.AI.chatbot.repository;
 
-import org.aspectj.weaver.patterns.TypePatternQuestions.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question, Long> {
+import com.AI.chatbot.model.Question;
+
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
+
 }
